+++ spec/integration/rest/api/v2010/account/sip/domain.spec.js
@@ -54,7 +54,8 @@ describe('Domain', function() {
                   'voice_status_callback_method': 'POST',
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-                  'emergency_calling_enabled': true
+                  'emergency_calling_enabled': true,
+                  'secure': true
               }
           ],
           'start': 0,
@@ -97,7 +98,8 @@ describe('Domain', function() {
                   'voice_status_callback_method': 'POST',
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-                  'emergency_calling_enabled': true
+                  'emergency_calling_enabled': true,
+                  'secure': true
               }
           ],
           'start': 0,
@@ -145,7 +147,8 @@ describe('Domain', function() {
                   'voice_status_callback_method': 'POST',
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-                  'emergency_calling_enabled': true
+                  'emergency_calling_enabled': true,
+                  'secure': true
               }
           ],
           'start': 0,
@@ -211,7 +214,8 @@ describe('Domain', function() {
                   'voice_status_callback_method': 'POST',
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-                  'emergency_calling_enabled': true
+                  'emergency_calling_enabled': true,
+                  'secure': true
               }
           ],
           'start': 0,
@@ -313,7 +317,8 @@ describe('Domain', function() {
           'voice_status_callback_method': 'POST',
           'voice_status_callback_url': null,
           'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-          'emergency_calling_enabled': true
+          'emergency_calling_enabled': true,
+          'secure': true
       };
 
       holodeck.mock(new Response(201, body));
@@ -377,7 +382,8 @@ describe('Domain', function() {
           'voice_status_callback_method': 'POST',
           'voice_status_callback_url': null,
           'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-          'emergency_calling_enabled': true
+          'emergency_calling_enabled': true,
+          'secure': true
       };
 
       holodeck.mock(new Response(200, body));
@@ -440,7 +446,8 @@ describe('Domain', function() {
           'voice_status_callback_method': 'POST',
           'voice_status_callback_url': null,
           'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
-          'emergency_calling_enabled': true
+          'emergency_calling_enabled': true,
+          'secure': true
       };
 
       holodeck.mock(new Response(200, body));