+++ CHANGES.md
@@ -1,6 +1,33 @@
 twilio-node changelog
 =====================
 
+[2020-04-01] Version 3.42.0
+---------------------------
+**Library - Fix**
+- [PR #551](https://github.com/twilio/twilio-node/pull/551): axios expects paramsSerializer not paramSerializer. Thanks to [@stephencaldwell](https://github.com/stephencaldwell)!
+- [PR #549](https://github.com/twilio/twilio-node/pull/549): refactor 'RestException' to properly extend 'Error'. Thanks to [@childish-sambino](https://github.com/childish-sambino)!
+
+**Api**
+- Add optional 'secure' parameter to SIP Domain
+
+**Authy**
+- Added an endpoint to list the challenges of a factor
+- Added optional parameter `Push` when updating a service to send the service level push factor configuration
+
+**Bulkexports**
+- exposing bulk exports (vault/slapchop) API as public beta API
+
+**Flex**
+- Adding `queue_stats_configuration` and `wfm_integrations` to Flex Configuration
+
+**Serverless**
+- Add Function Version Content endpoint
+- Allow build_sid to be optional for deployment requests
+
+**Supersim**
+- Remove `deactivated` status for Super SIM which is replaced by `inactive` **(breaking change)**
+
+
 [2020-03-18] Version 3.41.1
 ---------------------------
 **Library - Fix**